refactor: Modified to use a http client interface
Currently go-marathon uses net/http's client struct which is prohibiting the use of custom http clients. I have modified the library to use a client interfacde which will allow the passing in of custom interfaces. Currently this library is pointing at Udacity's fork of the eventsource package since this also had to be modified to accept interfaces.